+++ src/main/java/org/apache/aurora/scheduler/storage/durability/ThriftBackfill.java
@@ -23,7 +23,6 @@ import org.apache.aurora.GuavaUtils;
 import org.apache.aurora.gen.JobConfiguration;
 import org.apache.aurora.gen.JobUpdate;
 import org.apache.aurora.gen.JobUpdateInstructions;
-import org.apache.aurora.gen.Resource;
 import org.apache.aurora.gen.ResourceAggregate;
 import org.apache.aurora.gen.ScheduledTask;
 import org.apache.aurora.gen.TaskConfig;
@@ -41,10 +40,6 @@ import org.apache.aurora.scheduler.storage.entities.ITaskConfig;
 import static java.lang.String.format;
 import static java.util.Objects.requireNonNull;
 
-import static org.apache.aurora.scheduler.resources.ResourceType.CPUS;
-import static org.apache.aurora.scheduler.resources.ResourceType.DISK_MB;
-import static org.apache.aurora.scheduler.resources.ResourceType.RAM_MB;
-
 /**
  * Helps migrating thrift schema by populating deprecated and/or replacement fields.
  */
@@ -55,14 +50,6 @@ public final class ThriftBackfill {
   @Inject
   public ThriftBackfill(TierManager tierManager) {
     this.tierManager = requireNonNull(tierManager);
-  }
-
-  private static Resource getResource(Set<Resource> resources, ResourceType type) {
-    return resources.stream()
-            .filter(e -> ResourceType.fromResource(IResource.build(e)).equals(type))
-            .findFirst()
-            .orElseThrow(() ->
-                    new IllegalArgumentException("Missing resource definition for " + type));
   }
 
   /**
@@ -125,28 +112,16 @@ public final class ThriftBackfill {
    * @return Backfilled IResourceAggregate.
    */
   public static IResourceAggregate backfillResourceAggregate(ResourceAggregate aggregate) {
-    if (!aggregate.isSetResources() || aggregate.getResources().isEmpty()) {
-      aggregate.addToResources(Resource.numCpus(aggregate.getNumCpus()));
-      aggregate.addToResources(Resource.ramMb(aggregate.getRamMb()));
-      aggregate.addToResources(Resource.diskMb(aggregate.getDiskMb()));
-    } else {
-      EnumSet<ResourceType> quotaResources = QuotaManager.QUOTA_RESOURCE_TYPES;
-      if (aggregate.getResources().size() > quotaResources.size()) {
-        throw new IllegalArgumentException("Too many resource values in quota.");
-      }
+    EnumSet<ResourceType> quotaResources = QuotaManager.QUOTA_RESOURCE_TYPES;
+    if (aggregate.getResources().size() > quotaResources.size()) {
+      throw new IllegalArgumentException("Too many resource values in quota.");
+    }
 
-      if (!quotaResources.equals(aggregate.getResources().stream()
-              .map(e -> ResourceType.fromResource(IResource.build(e)))
-              .collect(Collectors.toSet()))) {
+    if (!quotaResources.equals(aggregate.getResources().stream()
+        .map(e -> ResourceType.fromResource(IResource.build(e)))
+        .collect(Collectors.toSet()))) {
 
-        throw new IllegalArgumentException("Quota resources must be exactly: " + quotaResources);
-      }
-      aggregate.setNumCpus(
-              getResource(aggregate.getResources(), CPUS).getNumCpus());
-      aggregate.setRamMb(
-              getResource(aggregate.getResources(), RAM_MB).getRamMb());
-      aggregate.setDiskMb(
-              getResource(aggregate.getResources(), DISK_MB).getDiskMb());
+      throw new IllegalArgumentException("Quota resources must be exactly: " + quotaResources);
     }
     return IResourceAggregate.build(aggregate);
   }