+++ src/main/python/apache/aurora/config/resource.py
@@ -85,7 +85,7 @@ class ResourceManager(object):
 
   @classmethod
   def resource_details_from_quota(cls, quota):
-    return cls.resource_details(cls._backfill_resources(quota))
+    return cls.resource_details(quota.resources)
 
   @classmethod
   def resource_details_from_task(cls, task):